ducky: change the balanced partition criteria
The assessment whether partitions are balanced were done by comparing
the number of partitions in each node with the average target ±20%. This
way went broken with the introduction of partition balancing domains.
The criteria is changed by this commit to make sure that the number of
partitions across noded is levelled within a scope of each domain
separately. Levelled means that min and max # of replicas differ by 1
at most.
Re #7418