txn: fix USEs on add partition & group
Before we introduced fetch_tx the state of the ongoing tx lived only
in memory so RP was querying only the stm's transient memory and not
the log. Now RP may rehydrate the state from the previous leader and
correct a tx's etag by writing to log. As a result we should check
not only the transient memory and the state rebuild from the log too

Fixes https://github.com/redpanda-data/redpanda-jepsen/issues/23