cluster/topic_table: fix quadratic behavior in notify_waiters
From issue 7665; Consider when recovering a large topic with many deltas and when _waiters is empty.
The deltas are applied from the log one-by-one and notify_waiters is called after each addition.
Then the copy_if will traverse the entire pending deltas every time, so we have O(n^2) behavior in the number of deltas.

This commit reduces the behavior to O(n) by storing the last sent index
of _pending_deltas