c/topics_frontend: validate partition condition before replicating
Controller uses optimistic approach to provide correctness when
executing commands i.e. result of the operation is determined when
records are applied to the state machine. This cause a problem when
an operation is retried mulitple times as it may cause the controller
log to grow really fast. Added preliminary validation to topics frontend
to prevent controller log from growing.

Signed-off-by: Michal Maslanka <michal@redpanda.com>