tests: improve package search to ignore releases with no tarball
This can happen during releases, when the release exists in github
but the package doesn't, either because the release is in flight or
because something went wrong.

We should not fail all our CI runs when the release is in this state:
let it skip back up to 2 versions to find one that has packages
available for upgrade tests.