cloud_storage: checks before closing remote segment
When a remote segment is closed, a flag is set. If there is another call
to close the segment, the flag prevents attempting to close resources in
the segment twice, by checking the flag and performing an early return.

Another option would be to just protect closing of the segment gate with
the flag, allowing the rest of the close method to run as before, but
this becomes hard to manage if the code in stop() grows with more
resources.