Avoid unnecessary move in topic_table::apply
In that method we created a deltas container and emplaced values
into it, then moved this value to the end of the pending deltas
member. This results in an extra move versus simply emplacing
the deltas directly in the member.

The deltas are non-trivial objects and the moves are in some
cases relatively expensive (since e.g., they have to copy all the bytes
of ss::string which has a small-string optimization).

Instead, just emplace the values directly in the pending
deltas array. There are no suspension points so this is safe and
equivalent to the old behavior.

Fixes #7669.