transactions: Lower tx transfering log severity
Not being able to transfer a transaction during leadership
transfer is not always a concerning error because the producer
may have finished its transactions and is already idle and doesn't
care about the session anymore.

As long as the producer has not timed out on the broker, a transaction
with next seq_number remains in ready/ongoing state. There is no
explicit API provided by the client to cleanup a producer [session]
state.