cluster/topic_table: refactor delta_cb_t to remove copy
This changes the delta_cb_t to take a std::span instead of a reference
to a vector. The new interface allows for us to avoid copying a subset
of `_pending_deltas` before calling all of the delta callbacks.

Note that this change assumes that offsets in `_pending_deltas` are
monotomically increasing from `_pending_deltas.begin()` to
`_pending_deltas.end()`.