c/members_manager: fixed assigning node_id when seed server was replaced
Fixed node id assignment logic to prevent the same node id from being
assigned twice. The node assignment logic was broken and it might lead
to situation in which members manager assigned the same node id twice.
Consider situation in which we initialize a 5 nodes cluster providing
all servers in seed servers list. When one of the nodes is
decommissioned it should be removed from the seed server list. In this
case the assignment logic will not skip over the last seed server id and
will reuse the same id twice.

Signed-off-by: Michal Maslanka <michal@redpanda.com>