kafka: fix reading aborted transactions on recovered partitions
aborted_transactions and make_reader were using different rules
on when to use remote vs. local partition.

In the bug case seen in #7043, we were reading from the remote
partition, but then abort_transactions was hitting the local
partition.

Fix this by modifying the condition in aborted_transactions
to use the same kafka offset comparison that is used in make_reader.

Fixes: https://github.com/redpanda-data/redpanda/issues/7043