archival: set test reconciliation interval
Previously, the archiver reconciliation interval was default initialised
to 0ms. This led to race conditions in tests that called directly into
the archiver (e.g. test_reconciliation_drop_ntp) as, generally, it's not
safe to call into the scheduler service concurrently.

This commit set the archiver reconciliation interval to be longer than the
duration of any sensible unit test.