tests/nodes_decommissioning_test: do not wait for partitions moving
Nodes decommissioning test was waiting for partition to begin moving. In
some cases all the partitions might have been moved before the wait
triggered leading to test failure. Removed waiting for decommissioning
status and partitions moving in favor of waiting for the node to be
removed from the cluster which is the end result of decommissioning
operation.

Fixes: #7816

Signed-off-by: Michal Maslanka <michal@redpanda.com>