cloud_storage: filter stopped segments on trim
When trimming segments in materialized_segments, we should ignore
segments which are already stopped. This avoids the following issues:

* A stopped segment is offloaded by materialized_segments. If the remote
  partition is stopping at the same time, it is iterating over it's
segment map while an entry from the map is erased, causing an iterator
invalidation.

* A segment is marked for eviction and later attempted to stop, but it
  has already been stopped by the partition. This is avoided now due to
checks inside the segment, but if the segment is filtered out early we
avoid this trap later.