cluster: initialize node ID assignment counter
The counter isn't currently initialized with any of the node IDs that
are assigned to seed servers. This can result in a seed server's node ID
being reused after it's been decommissioned.

This commit adds a layer of protection by initializing the node ID
counter to just past the highest assigned node ID.

I updated some of the existing tests with a stricter check that would
have caught this issue.

Another approach to protect this case is to check the removed set of
nodes when assigning a node ID. A separate change will add that.