cloud_storage: Fix race condition in materialized_segments
With large number of segments it is possible to run into the following
situation. When the partition is stopped we're waiting for the eviction
loop to process all segments scheduled for eviction. The list of evicted
segments is moved to the temporary variable and then the eviction loop
processes it.

The 'stop' method uses 'flush_evicted' method which checks the size of
the eviction list. The 'flush_evicted' method pushes eviction barrier to
the list only if the eviction list is not empty. But it is possible that
the 'flush_evicted' method will be called right after the eviction list
is moved to the temporary. In this case the 'flush_evicted' will return
early and wouldn't create a barrier. This will lead to a situation when
the 'remote_parititon' could be deleted before all segments stored in
the temporary eviction list are processed.

This commit splits the eviction list into two parts. The pending list
and in-flight list. New items are added to the pending list. Then the
eviction loop moves it to the in-flight list and processes it. The items
in the in-flight list are processed in FIFO order and removed right
after processing. This is because the list might contain a lot of
segments from multiple partitions. In this case when the eviction loop
will evict the eviction barrier the in-flight list won't be cleared yet
(because there are other segments remaining). The eviciton of the
barrier will let the 'remote_partition::stop' method to exit and and
'remote_partition' to be deleted despite the fact that the evicted
'remote_segment' instances that belongs to it are not deleted yet. This
will trigger an assertion.

(cherry picked from commit 5871ad3d58036e639cd41c69958ece96a4a7b1fc)