cloud_storage: Limit number of materialized segments
Prevoiously, number of materialized segments in the 'remote_partition'
was limited by the 'topic_partitions_per_shard' value. The limit is
calculated as 'topic_partitions_per_shard' * 2. The default value for
'topic_partitions_per_shard' is 7000 which gives us 14K materialized
segments per shard. This is way too high. The idle time for
materialized segment is 60s which is enough time to materialize a lot of
segments if they're cached and not too big.

The 'remote_segment' instance is relatively heavyweight. It contains
materialized segment index which could be relatively large (several
kilobytes). If we have a lot of cached small segments in cache the
client could materialize very large number of segments consuming a lot
of memory.

This commit changes the default to 'cache-size/segment-size' value. E.g.
if the cache size is 10GiB and segment size is 512MiB we will be able to
materialize 20 segments.