lsan: add lsan suppressions file
A memory leak pops up in our CI from time to time
(https://github.com/redpanda-data/redpanda/issues/3338). It is caused by
a very narrow race within Seastar when incoming connections arrive while
a socket server is being shutdown.

I've posted a "fix" for this in upstream Seastar (link below) which is
effectively rejected because it doesn't really fix the leak, but merely
cleans up on exit to make LSAN happy. AFAICT this leak can only happen
when a socket server is being shut down, and in Redpanda, this only
happens when Redpanda is shutting down. In this case it is harmless,
but annoying because it will occassionally fail our CI.

Adding an LSAN suppressions file allows us to instruct LSAN to ignore
this particular leak when it occassionally does trigger.

Fixes: https://github.com/redpanda-data/redpanda/issues/3338

See https://github.com/scylladb/seastar/pull/1265 for more discussion
and a link to an issue with a reproducer.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 778704b2177b125946608cbb206e11ac16602439)