cloud_storage: Evict segments based on heuristic
If the client is doing batch job the segments are scanned one by one and
never reused by the same client. Previously, the remote_partition didn't
take client behavior into account and this resulted in materialization
of large number of partitions.

To avoid this this commit adds a heuristic which is invoked when the
reader is complted scanning a segment. In this case the, if the reader
is the only user of the segment the segment will be evicted without
waiting 60s (our hardcoded idle timeout).