dev_cluster: graceful shutdown for ctrl-c
also removes the help menu. running the executor and reading from stdin
is proving hard to cancel nicely. i'm sure there is an easy way, but i'm
not able to figure it out. in any case, we have predictable ips so it
doesn't provide a lot of benefit at the moment compared to the improved
non-hanging ctrl-c behavior.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>