storage: clean up after aborted compaction
If the generation_id is updated while we write the compaction output, we
end up returning early without keeping track of the staging files. This
could result in files being left over, even after removal of the
partition.

This commit addresses this by immediately removing files that may go
unused upon exiting early out of a compaction due to a generation ID
mismatch.