storage: clean up staging files on deletion
If the generation_id is updated while we write the compaction output, we
end up returning early without keeping track of the staging files. This
could result in files being left over, even after removal of the
partition since we currently don't allow removing the NTP directory
while any unexpected files exist.

This commit addresses this by removing all files suffixed with
".staging" when a partition is deleted.

I considered an alternate fix wherein we kept track of all staging files
while compacting, but opted to scrap the approach, as it became a fairly
invasive change with several edge cases (e.g. staging files when
compacting a staged segment), and this fix will likely need to be
backported, so a simpler approach is preferrable.