kafka: track consumer groups subscriptions
In a few features like offset deletion and offset retention GC the
set of subscribed consumer topics are checked by the broker. However,
this set of topics is encoded in binary data produced by the client and
normally never inspected by the server.

However, for the well-known "consumer" type of consumer group the format
is standardized and the broker can decode this data to track the
subscriptions.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>