cloud_storage: reduce I/O during cache put
Seastar's recursive_touch_directory is at a minimum
two system calls (mkdir and fsync).  The vast majority
of the time, a partition's directory already exists,
so it is worth checking this.

We might want to improve this further by proceeding
to try and open the file immediately, and only
checking directory existence if that gives an error.