cloud_storage: Add transactions to segment_meta_*
Add transactions to segment_meta_column and segment_meta_column_frame.
The append_tx method optionally returns transaction if the operation
can't be performed without allocating memory. In case if the method
returned the transaction it has to be committed to take effect.

This was implemented with the following use case in mind. When we have a
column store with multiple columns and we need to add value to every
column the append operation may throw. The column store will end up
being in a corrupted state. To avoid this the append_tx method could be
used. If all append_tx calls are successulf the operation can be
completed safely by committing all transaction which will be safe
because commit never throws.