Replace method_id by method_info
Prior to this change we pass a method_id, which is a 32-bit integer,
down into the guts of the rpc send methods. However, in some cases
we would like to log the name of the method, and the ID is not useful
for this purpose.

In this change we swap out method_id for method_info, which bundles the
ID and a human-readable method name like "controller::join", with the
intent that the name can be used for logging or other identification
purposes.

This does not change the RPC call itself in any way.

Issue #7853.