cloud_storage: adds list api and put-object
The remote utility is extended to add support for list api and put
object. The list api is present in the cloud_storage clients, it is
exposed via remote here. The implementation allows for repeated calls to
the list api if the results cannot be fetched fully in a single call.

The put object API added is suitable for small payloads as it uses a
simple string rather than a stream. It is implemented to support
creating lock files for topic recovery.