s3: Check failbit after XML formatting
The XML request is generated using std::stringstream. In case if
bad_alloc is thrown during the formatting the stringstream won't
propagate the exception and the content of the stream will be invalid.
We need to check failbit and throw exception if formatting failed.
Otherwise the code will send bad XML request.