tests: deflake partition_balancer_test.test_decommission
When a broker kill command races with applying a new config delta,
partial configuration file is written to disk and the subsequent
start fails. This is a known issue in the current implementation.

```
try {
        // TODO: do a clean write tmp + mv, to
        // avoid possibility of torn writes.  Note
        // that an fsync is not necessary as long
        // as the write is atomic: it is safe for
        // the cache to be a little behind.
        iobuf buf;
        buf.append(out.c_str(), out.size());
        co_await write_fully(cache_path(), std::move(buf));
    } catch (...) {
        // Failure to write the cache (maybe we're
        // out of disk space?) is not fatal.  It
        // just means that on next restart we'll
        // apply the deltas since the last time we
        // successfully wrote it.
        vlog(
          clusterlog.warn,
          "Failed writing config cache: {}",
          std::current_exception());
    }
```

In this test, when a NodeStopper injects the kill failure, it is racing
with application of `raft_learner_recovery_rate` configuration. This
patch doesn't fix the underlying root cause but rather adapts the test
to this behavior by waiting for the config_version to be applied on the
nodes that are not killed.