utils: Implement random access without decompression
Skip method is added to deltafor_decoder. The method allows the decoder
to be moved to particular row without decompressing all previous rows.
To be able to do this we need to save the deltafor_stream_pos_t object
in advance. The 'pos' object points to a particular batch in the encoded
stream and can be generated by the deltafor_encoder. The idea is that
the user of the encoder can cache this 'pos' objects and use them to
implement quick random access inside the decoded stream.