cloud_storage: Optimize column random access
Use deltafor_stream_pos_t to speed up index lookups. The problem is that
to materialize the segment_meta instance we need to read each field from
the corresponding column. We have 12 fields so there are 12 random read
operations on 12 columns. Normally, this would require scanning up to
max_frame_size elements per column which is 768 rows in worst case. In
order to avoid doing this the 'skip' method of the deltafor_decoder is
used to skip unused rows.

The 'hints' for read operations are supposed to be cached by the column
store. The total size of all hints is acceptable even if we will store a
hint per row per column.