cloud_storage: Implement basic column_store
Implement column store and materialization operations. The column store
contains 12 columns. One column per field in segment meta struct. It
also contains a btree map that maps from index in the column store to
the vector of hints. Hint is an instance of deltafor_stream_pos_t struct
generated using the column. The hints are used to speedup the index
lookups which are used during materialization of the segment_meta
objects.

Only 1 out of 8 rows is indexed so the overal memory overhead of
indexing is low. Materialization step uses index to fetch values from
every column so it has to use a 'hint' to optimize this and to avoid
decoding the stream prior to a value.