Introduce timeout_spec to replace time_point
Currently we generally use a single time_point to indicate timeouts, but
this loses the information about what the duration of the timeout was.

Introduce timeout_spec which packs a time_point representing the
timeout point (as before) and also the duration associated with this
timeout. The duration can be used for improved logging and diagnosis
when a timeout occurs.

This change only introduces timeout_spec in the lower level rpc layer
and a few related places but does not yet make a wholesale change from
raw time_point usage treewide.

No change in behavior is expected/intended in this change.

Issue #7853.

(cherry picked from commit e5027b1920f4e470c352af499f10c3ab7be9e23d)