archival: Do not run upload/sync/housekeeping loops ...
util we have stable leadership. The code was using is_elected_leader. In
this state the underlying raft group is not guaranteed to be able to
replicate the message. All this internal background loops are actually
replicating messages and can fail if the leadership is not stable.

The commit fixes this by using 'cluster::partition::is_leader' instaead
of using 'cluster::partition::is_elected_leader'.

This change is based on commit 3334790cd3078152a2f9ec0d28ed20f69970151e
with some differences. The important one is that in v22.3.x we still
have archival::scheduler_service which uses is_elected_leader to decide
which archivers to create. This has to be chaged, otherwise the
archivers might be created too early which will prevent them from
functioning properly (the upload loop will immediately exit since it
uses different condition  now).