Add a flush indicator to batched_output_stream
The write() method on batched_output_stream may or may not
flush the underlying stream, depending on "conditions".

In some cases, the caller might like to know whether a flush occurred
or not, for example because a lack of flushing is one reason an RPC
may sit in the in-memory buffer rather than being sent immediately.

This change returns a boolean indicating whether a flush occurred in
write. Some callers have to be modified to ignore this value.

Issue #7853.