Augment RPC timeout logging
We augment the RPC timeout logging to include the name of the RPC
method that failed, as well as more detailed information about the
timing:

 - The configured duration of the timeout
 - The current time since the timeout start point
 - The times at which the request was enqueued, dispatched and written

We also include whether the request was flushed by the batched output
stream or not.

Issue #7853.

(cherry picked from commit 30748dfa368b402671c3ed570c42cf2095394660)