cloud_storage: fix kafka::offset lookup in partition_manifest
When serving a fetch request by a kafka::offset, we previously used the
`partition_manifest` to perform a segment lookup by
`kafka::offset_cast(ko)` first in order to traverse the segments forward
to find the actual segment that contained `ko`.

This doesn't work when the manifest has been truncated, e.g. if the
casted offset falls before the start of the manifest, we would
previously return that no segment exists for the fetch. This could
result in segments erronesouly not being returned, and fetches
erroneously being met with no data when some existed.

This commit fixes the behavior to no longer use the casted segment
lookup and adds some test coverage for kafka::offset lookups.

Fixes #7708