c/topic_table: don't store replica revisions in the update struct
Previosly during an update we stored both the current revisions map
and the original one. We also updated the current assignment in case
of cancellation, which resulted in the target map being lost. But
target revision map is easily calculated from the old map and the
assignments themselves. So we need to store only the old revisions map.