features: fix populating original_version on upgrade
Since we added snapshots, older redpanda nodes will skip
updates from before their snapshot offset.  This means
they'll miss opportunities to learn their original version.

We need a special case to pay attention to updating the
original version if it is unset, even if the update
we're replaying is before the snapshot offset.