rptest: tweak SISettings to allow using ABS
This commit modifies 'RedpandaService' to use different cloud storage
providers based on the 'cloud_storage_type' parameter injected in the
test context.

The logical starting point for this commit is the change to SISettings.
It now acceps a 'TestContext' as it's first argument and inspects in
order to determine if it should use S3 or ABS for the current test.