kafka: track offset commit/expire in memory
This updates kafka::offset_metadata to contain commit and expire
timestamps that reflect metadata associated with group offsets which are
used in offset retention policy.

Prior to this commit the timestamps were present in the consumer group
on disk metadata and also arrived via kafka client requests, but in both
cases, we threw this information away and didn't represent in memory
because there were no users of the information.

There are two paths that populate the in-memory state. The first is from
kafka client via offset commit request:

  store_offsets(offset_commit_request)
  store_txn_offsets(txn_offset_commit_request)

    - in both cases the commit/expire timestamps arrive from the client

The second path to populate the in memory state is via recovery:

  group_manager::do_recover_group

    - in this case the commit/expire offsets are present in the on-disk
      state that is recovery, and are passed through to the in-memory
      state.

  group_stm::update_prepared

    - there does not appear to be any source for commit/expire offset
      information. this may be due to a difference in our txn
      implementations.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>