k/group: bump error log messages to debug
Currently to be able to see context behind congroup related errors,
kafka logger needs to be set to trace level, which is really polluted
with lines like
```
kafka - fetch.cc:348 - fetch_ntps_in_parallel: for 1 partitions returning 0 total bytes
```
The idea behind this change was that we don't need to do any tuning to be
able to see anomalies in the log. However these cases are parts of normal
operation and return clear error codes, so the bump is limited to debug.