kafka/protocol: Dont gen opt block for []primitive
- In situations where a kafka schema file contains a type such as []int,
an array of any primitive type, a latent bug exists that when the
`nullableVersions` field is tagged onto it, compiler errors occur. The
compile error is valid because arrays do not contain null elements.

- The error is due to the fact that the generator generates code which
attempts to deference a POD type.

- The solution is just to remove the portion of the template that
performs this. Running diff on all of the generated code before and
after this commit shows no changes. Furthermore analyzing why it makes
sense reinforces why it should not exist. It occurs in a condition where
`field.is_array` is true and `field.type().value_type().is_struct` is
false, where when put together means the case of generating a block of
code for a primitive array has occurred. In this situation there is some
abiguity to `nullableVersions`, i.e. is it referring to the outer
type (the array) or the values within the array `[](optional<int>)`?
Sticking with the assumption that these tag attributes apply to the
field itself (array in this case), removing this block of code removes
the ability for any type of optional to be generated when the type is of
an array of primitves.