cloud_storage: generalize abort source handling in client pool
Rather than having remote subscribe to an abort source
to shutdown the client after leasing it, require that
anyone leasing a client object provide an abort source,
which the client will subscribe to.

As with the change to retry_chain_node, there are no
use cases where a proper abort source is not available,
so this is not an onerous interface change.

The abort source has to outlive the lease, but that's
a common pattern in our code, where abort sources
are passed by reference from longer lived objects
into shorted lived objects.