tests: stabilize controller leader after config restart
In KgoVerifierWithSiTest, we sometimes use a set_cluster_configs
immediately before an alter_topic_config.

There was a hidden bug here, which was un-hidden when error
handling on Rpk wrapper was fixed: these alter_topic_config
calls were not working, because the controller had no leader
immediately after the config-driven restart.

This caused failures in
- KgoVerifierWithSiTestLargeSegments.test_si_with_timeboxed
- KgoVerifierWithSiTestSmallSegments.test_si_with_timeboxed