tests: add scaled down version of cloud_retention_test
scale_tests/cloud_retention_test has some useful checks that help ensure
we exercise codepaths like fetching from before the start of the
manifest. In #8110, this test caught a real bug and is worth running in
CI. This commit copies the existing test but tunes it down to be able to
run in with docker.