rm_stm: fix use-after-move
try_emplace does not move from its value parameter when an insertion
does not take place. however, a change in

   https://github.com/redpanda-data/redpanda/commit/898e54132791c2fb591952c806ba1983da56ec08#diff-31c312a88efb2dcf54d8fc5447d59316554d9f7df36fa3125cacbf768a3bdeb0R2356-R2358

moved the value into a wrapper structure creating a temporary which
would not have been moved from if insertion didn't take place, but the
entry was still moved from.

the consequence of this is that an empty entry may be inserted into the
sequence table during an upsert.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>