k/topic_utils: do not return an error when waiting for leaders failed
Recently introduced change (waiting for leaders before returning from
create topics) change the behavior of `CreateTopicsRequest` handler.
Previously handler was waiting for topic partitions to be created but
even if that timed out it returned success to the client. Waiting for
leaders threw an exception when timeout was trigger which caused client
connection to be dropped.

Fixed an error by going back to previous behavior i.e. ignoring waiting
timeouts. The timeout does not indicate that topic creation failed
actually when waiting for creation topic already exists.

Fixes: #7942

Signed-off-by: Michal Maslanka <michal@redpanda.com>