controller: include logical version in bootstrap record
This enables nodes reading the record to fast-forward their
feature table to this version, without waiting for the feature_manager
logic to see all nodes' versions and activate that way.

Usually this is superfluous, when a founding triplet of nodes
will be up promptly, but in some cases we might run with just
two nodes for some time, where these two nodes would otherwise
wait for the third to report its version before advancing
the feature table version.