cluster: set feature_table version during bootstrap
In some circumstances, a cluster can have two nodes up+bootstrapped,
with an active controller raft group, but a third founding node
trying to join the cluster by requesting an ID (because this node
reported in earlier, but had trouble following through with the
bootstrap process).

Because handing out node IDs is behind a feature gate, this
would fail, and the cluster logical version would not advance
because there was a node in the members_table who had not
reported its logical version.

Avoid this class of issues by including logical version in
the controller bootstrap event, so that as soon as there is
a bootstrap message in the log, we are certain to have activated
all the features that a node might need to proceed.

Fixes https://github.com/redpanda-data/redpanda/issues/8203

(cherry picked from commit 54120c9e1f982a36963749ea15799eaa9b43c11c)