rpk: add namespaced resource to k8s debug bundle
This issues request to the REST k8s API and saves
the raw json response.

In the future, depending on how we implement the
RBAC in the helm chart, we can use the discovery
client to list the resources in the cluster and
list them all.
(similar to kubectl get api-resources')