cloud_storage,cluster: Add topic recovery service
The topic recovery service and its related RPC components are added. The
topic recovery service is a sharded service which starts a background
recovery task on calling its API. The service is only expected to be
active on one shard per cluster at a time. The location of this shard
is the same as controller leader, but before starting recovery, all
other nodes are queried to check if a recovery is not running on them.
This is to ensure that a recovery (which is a long running process) is
not running on a node which was a controller leader but lost
leadership. The service propagates its status on all shards on the node
where it is running. This status is used to prevent duplicate instances
of recovery.

The service performs the following actions per recovery:

* Scan the configured bucket for topic manifest JSON files.
* Download and parse the manifests. Exclude topics which exist in
  cluster.
* If the request contained a pattern for topic names, match and filter
  topic manifests against this pattern.
* Create topics using the parsed replication factors and partition
  replicas. Cache the retention settings in manifest, and use a default
  local retention size of 1 GiB.
* Start a background task to periodically scan the bucket for results of
  the topic download process. The partition download manager in a
  subsequent change will upload the results of download to the bucket.
* Finally once all the downloads are complete, either successful or
  failed, reset the topic settings to the retention from the manifests
  and exit.

 The RPC related changes are helpers to be able to query if recovery is
 running on another node.