cloud_storage,admin: Wire in topic recovery service
The topic recovery service is added to the admin server. An endpoint is
added to the admin server to respond to API calls to initiate topic
recovery.

The topic recovery service and status frontend are added to the
partition recovery manager. This enables partition recovery to upload
its results to the S3 bucket where the topic recovery service can
periodically update its status based on results seen.

The topic recovery service and frontend are added to the application and
the dependencies are wired together.

All of the above changes are interlinked and need to be introduced
together.

The shadow indexing API endpoints are renamed to cloud storage.