tests/fixture: Allow redpanda fixture to use custom port
The redpanda fixture has a hardcoded port number for shadow indexing
config. With the changes in the unit tests which run tests on different
ports for different modules, it is necessary to be able to override the
port so that the fixture class using redpanda fixture is able to use
its own port for the imposter.