c/members_backend: simplified and fixed tracking of last decommission
Members backend was tracking last node decommission revision to be able
to cancel all related partition movements. The tracking was broken as it
might be the case that the revision map was updated by the next
decommission update while the previous recommission was still being
processed. In order to fix this issue and simplify tracking of last
decommission revision id introduced tracking previous decommission
revision inside of recommission update metadata object. This way a
certain recommission action is always related with correct decommission
revision.

Fixes: #8218

Signed-off-by: Michal Maslanka <michal@redpanda.com>