tests/offset_for_leader_epoch: allow gaps in term space
In redpanda not every term increase result in a leader being elected.
The test assumed that last offset for term `1` should always be preset
but it may not be the case when election in term `1` was inconclusive.

Made the assertion in test aware of the fact that there may be a gap in
term space.

Fixes: #7397

Signed-off-by: Michal Maslanka <michal@redpanda.com>