cloud_storage: Add retention for recovered topics
A new property called
cloud_storage_recovery_temporary_retention_bytes_default is added which
is used to drive the local retention for newly created topics. This
bound is active for the period of topic recovery. Once this recovery is
complete, the local retention size is reset to the default value for
cluster. The value is set to 1 GiB by default.