features: add bootstrap_active_version
The "normal" set_active_version method increments the version but does
not activate features (though it may make them available).  This is
because activating features on upgrade is optional, and that optionality
is respected at time of writing to the controller log, not replaying it.

The reason for making auto-activation optional is to enable cautious
upgrades, but during cluster bootstrap this motivation goes away, so
we will unconditionally switch on all the available features when we
see the bootstrap message's cluster version in the controller log.  That
is what this function does, as well as calling through to set_active_version.

(cherry picked from commit 98ba9430855e1d8b0db09b9b1d18cdc16fe3b1e1)