test: fix handling of bool in cluster_config_test
the flipped value for testing config updates to boolean options comes
from the 'example' field on a property which for boolean properties is
!_default and is populated on the redpanda side in the json repr.

this means that no boolean options were using the boolean type case.
however, group_offset_retention_enabled it an optional<bool> so it
wasn't annotated with an example field inside redpanda.

since bools are handled by populating `updates` with native types, and
the expected vs found checks handle bools as strings, the test was
failing the assertion `true == True`.

this commit handles the case where valid_value is native bool type and
converted the expected type to string if the actual type is a bool, too.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>