cloud_roles: lift azure storage API version
This commit lifts the hardcoded storage API version to a more proeminent
location. We hardcode it in order to make sure that the version we've
tested against is used in the field. These versions have a long
lifetime ('2014.*' is still available), so there's no risk of it simply
getting removed.