Serialize --io-properties as json
The --io-properties flag to redpanda takes a YAML input, but having
newlines in the command line is inconvenient for a variety of reasons,
e.g., it breaks the assumption that log lines have 1 line, it requires
careful re-escaping if you want to use the command line yourself, etc.

Instead, serialize it as a single json line, since json is valid YAML.

Issue #8283.